General improvements to site
- Add php templating
- Tidy CSS files, remove spaces/empty lines
- Convert old float grid with flex box, see notes
- Changed template to match new php style

NOTE - usually the grid system would be slightly different, but I have reduced the percent amount which creates a 'gutter' between the columns and then used `justify-content` to evenly space the columns. In the old grid this was done with margins so I have tried to match the style so its not a huge change for you to do.

If you wanna check this first you can checkout the branch  by doing `git fetch && git checkout origin/general`, which will put you in a preview mode, to exit just do `git checkout gh-pages` to switch back to your local branch.

This can wait until Im with you, but if you take a back up before doing it, then you should be fine, worse case is you revert back to an earlier commit.